bgpd: Change unsigned short to size_t for storing the length of as-path string
Avoid truncation at 65k, while aspath_make_str_count() uses "int len" then
assigns to aspath->str_len.

+++ bgpd/bgp_aspath.c
@@ -698,7 +698,7 @@ struct aspath *aspath_intern(struct aspath *aspath)
    reference count and AS path string is cleared. */
 struct aspath *aspath_dup(struct aspath *aspath)
 {
-	unsigned short buflen = aspath->str_len + 1;
+	size_t buflen = aspath->str_len + 1;
 	struct aspath *new;
 
 	new = XCALLOC(MTYPE_AS_PATH, sizeof(struct aspath));
